Add CrypticBinaryBlockIterator

+++ core/src/main/kotlin/net/plcarmel/encryptedsequences/core/encryption/definitions/FixedSizeWordEncryptionAlgo.kt
@@ -2,6 +2,6 @@ package net.plcarmel.encryptedsequences.core.encryption.definitions
 
 interface FixedSizeWordEncryptionAlgo : FixedSizeWordProcessor {
 
-  fun encrypt(word: IntArray, at: Int = 0)
+  fun encrypt(word: ByteArray, at: Int = 0)
 
 }
